refactor: minimize fumadocs integration diff
