+++ web/next/content/docs/manage/theming.mdx
@@ -9,7 +9,7 @@ ZeroStarter uses [Tailwind CSS v4](https://tailwindcss.com) with CSS custom prop
 
 ## Dark Mode
 
-Dark mode is handled globally by `next-themes` with system preference detection:
+Dark mode is handled by `next-themes` with system preference detection:
 
 ```ts
 // web/next/src/app/providers.tsx
@@ -19,14 +19,6 @@ Dark mode is handled globally by `next-themes` with system preference detection:
 ```
 
 The theme toggle component at `web/next/src/components/mode-toggle.tsx` cycles through: **system → light → dark → system**.
-
-Docs and blog pages still use Fumadocs `RootProvider` directly at the route level, with Fumadocs' internal theme provider disabled:
-
-```ts
-<RootProvider theme={{ enabled: false }}>{children}</RootProvider>
-```
-
-That keeps a single source of truth for theme state while still enabling Fumadocs layout contexts on the content routes. The docs route adds its own local search provider on top of that because the sidebar search trigger is custom.
 
 ## CSS Variables
