+++ web/next/src/app/(content)/blog/[[...slug]]/page.tsx
@@ -8,19 +8,17 @@ import {
 } from "@/lib/fumadocs"
 import { blogSource } from "@/lib/source"
 
-export default async function Page(props: { params: Promise<{ slug: string[] }> }) {
-  const params = await props.params
-  const pageData = await getPageData(params, blogSource)
+export default async function Page(props: { params: Promise<{ slug?: string[] }> }) {
+  const pageData = await getPageData(props.params, blogSource)
   return renderPageContent(pageData)
 }
 
-export const generateStaticParams = createGenerateStaticParams(blogSource, { includeRoot: false })
+export const generateStaticParams = createGenerateStaticParams(blogSource)
 
 export async function generateMetadata(props: {
-  params: Promise<{ slug: string[] }>
+  params: Promise<{ slug?: string[] }>
 }): Promise<Metadata> {
-  const params = await props.params
-  return generatePageMetadata(params, {
+  return generatePageMetadata(props.params, {
     source: blogSource,
     ogPath: "/api/og/blog",
     ogType: "article",
